Keep thinking about the interactions function
1. First important thing I noted, just merely assigning the board to a new variable temp_board does NOT copy the board. After I change the
board temp_board changes too because they both point to the same location in memory. I need to use deepcopy for this.
2. Since I think there will be a lot of lines of code I tried to be clever and create another method which I will call within the the
interactions method. This took me a lot time because I had a misunderstanding about Python classes. Thing to remember: if I want to call a
method cellDeath within another class method I need to write self.cellDeath()! The parameter self refers to a particular instance of the
class. So self.board is actually an instance variable (and I thought it was a class variable). Having said that you can still use self.board
throughout the class I believe and it will change accordingly.