Finally sorted out cell death
I added all the differnt cases for the interactions method for cell death. Initially I thought that I did not have to treat the first row
and first column seperately as it was not running into errors. Then after doing some test cases (with only 1 cell in top row and some in
the bottom row I was getting a non zero neighbour count for the first cell which was strange!!) I realised that this is because if we have
i=0 and we look at board[i-1] we have board[-1] and that is an element in the last row! So I treated all cases seperately, ie all the
corners, first and last rows and columns and the rest (middle).