File for drawing a grid
Finally I tried to tackle the issue of how to represent the board. I googled this question. I first came up across Turtle but thought this
could be a little tricky still so I found a better library called Pillow/PIL (Python Imagining Library). It did not take long to learn how
to draw lines and spheres. Then just have to loop through the list and whenever there is a 1 create a sphere (cell) in that positon to
create the initial board.