New method which extends the board by one layer
I wrote a new method which appends an extra layer of zeros. So if array is say 3x3 initially this will go through each row and add a 0
on both ends and then create a row previous to first and after last which contains len + 2 or 5 0s. So the resulting array is 5x5.