New method
Added a method which checks if the board needs to be extended, that is when there are at least 3 cells (1s) in a row as then a new cell
will be created in a yet to be made row/column. It does this be going through each of the 4 edges and keeps a count of 1s, resetting the
count to 0 if it encounters a 0. It also keeps checking if count = 3 since if there is just a single occurance like that we will want to
resize the board. We resize the board in such a way so it always a square grid