The game is working correctly I think!
So I tried a few test cases and seems like it is killing and creating cells in the right way and the board is expanding as necessary.
Perhaps I need to try to do it so the board is always the same size? Also the images appear in different windows and for more than 3
iterations it becomes confusing which one is first so need to think about improving this. Must be a way to name them easily. Or maybe
the image can close after each one just like I did in matplotlib.