Fixing a bug
When I looked at the spaceships test cases I saw on wikipedia I noticed a bug in my code. When there were 3 cells in a row on the previous
to outermost edge there would be no extra layers of dead cells created. But then I need to create a cell on the extra layer but never will
as my loop doesn't cover it, so instead I check the last two outer edge layers and if there is a 1 in any I will add extra layers.